requirements.txt
####### Generic Django Requirements #######
# For more informations about file format
# See https://pip.pypa.io/en/stable/reference/pip_install/#requirements-file-format
###### Requirements with Version Specifiers ######
# See https://www.python.org/dev/peps/pep-0440/#version-specifiers
# Django 3.0
Django >= 3.0.7, < 4.0
# Custom extensions for the Django Framework
django-extensions >= 2.2.9, < 3.0
# Improved debugger
Werkzeug >= 1.0.1, < 2.0
# 12factor .env file for settings.py
Django-environ >= 0.4.5, < 1.0
# REST api @INSTALLED_APP
djangorestframework >=3.11.0, < 4.0
# Django DRY forms
django-crispy-forms >= 1.9.1, < 2.0
crispy-forms-foundation >= 0.8.0, < 1.0
# Python-PostgreSQL Database Adapter
# psycopg2 >= 2.9, < 3.0
psycopg2-binary >= 2.9, < 3.0 # to avoid compiling
# Recommanded password Hasher
# django[argon2] >= 3.2.3, < 4.0
argon2-cffi >= 20.1.0, < 21.0 # to avoid compiling
# Django Debug Toolbar @INSTALLED_APP
django-debug-toolbar >= 3.0a1, < 4.0
# Python linter
pylint >= 2.5.3, < 3.0
# Django linter
pylint-django >= 2.0.15, < 3.0
# model scheme
graphviz >= 0.14, < 1.0
# static file management @MIDDLEWARE
whitenoise >= 5.1.0, < 6.0
# Python WSGI HTTP Server for UNIX
gunicorn >= 20.0.4, < 21.0